requirements.txt
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==3.7.1
async-generator==1.10
asyncpg==0.29.0
attrs==23.2.0
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email-validator==2.1.0.post1
fastapi==0.110.0
fastapi-login==1.9.3
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
iso8601==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
orjson==3.9.15
outcome==1.3.0.post0
passlib==1.7.4
pycparser==2.21
pydantic==2.6.2
pydantic-extra-types==2.5.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
PyJWT==2.8.0
pypika-tortoise==0.1.6
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
starlette==0.36.3
tortoise-orm==0.20.0
trio==0.21.0
typing_extensions==4.9.0
ujson==5.9.0
uvicorn==0.27.1
watchfiles==0.21.0
websockets==12.0
